requirements.txt
geopandas==1.0.1
matplotlib==3.9.2
numpy==2.1.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
rasterio==1.4.1
scikit_learn==1.5.2
Shapely==2.0.6
torch==2.5.0+cu118
ultralytics==8.3.23

